Move service annotations to separate package
