+++ annotations/src/main/java/net/spals/appbuilder/annotations/service/AutoBindProvider.java
@@ -1,11 +1,11 @@
-package net.spals.appbuilder.annotations;
+package net.spals.appbuilder.annotations.service;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import static net.spals.appbuilder.annotations.AutoBindProvider.ProviderScope.SINGLETON;
+import static net.spals.appbuilder.annotations.service.AutoBindProvider.ProviderScope.SINGLETON;
 
 /**
  * Marks a service provider to be auto-bound
