+++ app-test/src/test/java/net/spals/appbuilder/app/modules/AutoBindServicesModuleTest.java
@@ -6,9 +6,9 @@ import com.google.inject.*;
 import com.google.inject.binder.AnnotatedBindingBuilder;
 import com.google.inject.binder.ScopedBindingBuilder;
 import com.google.inject.servlet.ServletScopes;
-import net.spals.appbuilder.annotations.AutoBindInMap;
-import net.spals.appbuilder.annotations.AutoBindProvider;
-import net.spals.appbuilder.annotations.AutoBindSingleton;
+import net.spals.appbuilder.annotations.service.AutoBindInMap;
+import net.spals.appbuilder.annotations.service.AutoBindProvider;
+import net.spals.appbuilder.annotations.service.AutoBindSingleton;
 import net.spals.appbuilder.app.modules.AutoBindServicesModule.AutoBoundProvider;
 import org.hamcrest.Matchers;
 import org.reflections.Reflections;
@@ -18,7 +18,7 @@ import org.testng.annotations.Test;
 import java.util.List;
 
 import static com.googlecode.catchexception.CatchException.verifyException;
-import static net.spals.appbuilder.annotations.AutoBindProvider.ProviderScope.*;
+import static net.spals.appbuilder.annotations.service.AutoBindProvider.ProviderScope.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isA;
