+++ app/src/main/java/net/spals/appbuilder/app/modules/AutoBindServicesModule.java
@@ -7,8 +7,8 @@ import com.google.inject.multibindings.MapBinder;
 import com.google.inject.multibindings.Multibinder;
 import com.google.inject.servlet.ServletScopes;
 import jersey.repackaged.com.google.common.base.Throwables;
-import net.spals.appbuilder.annotations.*;
-import net.spals.appbuilder.annotations.AutoBindProvider.ProviderScope;
+import net.spals.appbuilder.annotations.service.*;
+import net.spals.appbuilder.annotations.service.AutoBindProvider.ProviderScope;
 import org.reflections.Reflections;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
